+++ api/callbacks.go
@@ -4,6 +4,15 @@ package api
 #include "bindings.h"
 
 // typedefs for _cgo functions
+typedef int64_t (*get_fn)(db_t *ptr, Buffer key, Buffer val);
+typedef void (*set_fn)(db_t *ptr, Buffer key, Buffer val);
+
+// forward declarations (db_cgo.go)
+int64_t cGet_cgo(db_t *ptr, Buffer key, Buffer val);
+void cSet_cgo(db_t *ptr, Buffer key, Buffer val);
+
+
+// typedefs for _cgo functions
 typedef int32_t (*human_address_fn)(api_t*, Buffer, Buffer);
 typedef int32_t (*canonical_address_fn)(api_t*, Buffer, Buffer);
 
@@ -14,6 +23,49 @@ int32_t cCanonicalAddress_cgo(api_t *ptr, Buffer human, Buffer canon);
 import "C"
 
 import "unsafe"
+
+// Note: we have to include all exports in the same file (at least since they both import bindings.h),
+// or get odd cgo build errors about duplicate definitions
+
+/****** DB ********/
+
+type KVStore interface {
+	Get(key []byte) []byte
+	Set(key, value []byte)
+}
+
+var db_vtable = C.DB_vtable{
+	c_get: (C.get_fn)(C.cGet_cgo),
+	c_set: (C.set_fn)(C.cSet_cgo),
+}
+
+func buildDB(kv KVStore) C.DB {
+	return C.DB{
+		state:  (*C.db_t)(unsafe.Pointer(&kv)),
+		vtable: db_vtable,
+	}
+}
+
+//export cGet
+func cGet(ptr *C.db_t, key C.Buffer, val C.Buffer) i64 {
+	kv := *(*KVStore)(unsafe.Pointer(ptr))
+	k := receiveSlice(key)
+	v := kv.Get(k)
+	if len(v) == 0 {
+		return 0
+	}
+	return writeToBuffer(val, v)
+}
+
+//export cSet
+func cSet(ptr *C.db_t, key C.Buffer, val C.Buffer) {
+	kv := *(*KVStore)(unsafe.Pointer(ptr))
+	k := receiveSlice(key)
+	v := receiveSlice(val)
+	kv.Set(k, v)
+}
+
+/***** GoAPI *******/
 
 type HumanAddress func([]byte) string
 type CanonicalAddress func(string) []byte
@@ -31,7 +83,7 @@ var api_vtable = C.GoApi_vtable{
 func buildAPI(api GoAPI) C.GoApi {
 	return C.GoApi{
 		state:  (*C.api_t)(unsafe.Pointer(&api)),
-		vtable: DBvtable,
+		vtable: api_vtable,
 	}
 }
 
@@ -43,7 +95,7 @@ func cHumanAddress(ptr *C.api_t, canon C.Buffer, human C.Buffer) i32 {
 	if len(h) == 0 {
 		return 0
 	}
-	return writeToBuffer(human, []byte(h))
+	return i32(writeToBuffer(human, []byte(h)))
 }
 
 //export cCanonicalAddress
@@ -54,5 +106,5 @@ func cCanonicalAddress(ptr *C.api_t, human C.Buffer, canon C.Buffer) i32 {
 	if len(c) == 0 {
 		return 0
 	}
-	return writeToBuffer(canon, c)
+	return i32(writeToBuffer(canon, c))
 }
